Fixed two problems associated with SAOSAC rays: dither and time
SAOSAC rays are already rolled (i.e., already in spacecraft
coordinates).  So do not apply roll with the 0 amplitude dither.

The new experimental SAOSAC ray files that contain dither do not start
at t=0, but some large MJDsecs.  This had to be subtracted off when
reading the rays.  This will have to be revisited for full support of
dithered SAOSAC rays.