flatten marx directory and convert to automake
This proved more difficult than expected; in particular to pick
the right inputs for libmarx.a (which is far more than the *lis file states).
Because everything depends on everything (the directory structure on configure.ac,
the configure.ac on automake, automake on my ability ot formulate the right rules, ...)
this is not a small atomic commit but rather a "dumb everything in one buckt".

At this point automake + autoconf + autoheader work and compile
(I did not test if marx runs, yet!).
There is no use of libtool yet.