Change defaults in marx.par for HRC PSF
An email from Vinay Kashyap summarizes the reason for this change:
We discovered a problem with how HRC raytraces of point sources are
handled.  See attached figure (figure removed because data is not public)
 of radial surface brightness
profiles of data from a suspected point source vs different methods of
raytraces.  Green is bare HRMA, sans detector; blue is data; black is
current default ChaRT+MARX; and red is ChaRT+MARX with different HRC-I
blur parameters (see below).  The curves are normalized to a peak of
1.

As you can see, the default raytrace predicts a significantly larger
wing than is observed.  We believe we have tracked this down to a
difference in how MARX renormalizes the amplitude of the 2D Lorentzian
compared to the Sherpa model function.

MARX uses a 2 Gaussian + 1 Lorentzian function to model the HRC
detector blur.  Diab tested cases where he manually set the amplitudes
of the second Gaussian and the Lorentzian to zero, and gets a
reasonably good match to the data --
HRC-I-BlurG1FWHM=12.44
HRC-I-BlurG2Amp=0
HRC-I-BlurL1Amp=0
The same changes would also apply to HRC-S-*.