Change data type of PI and PHA columns written to match CIAO
This is required to dmmerge marx2fits output to e.g. ACIS stowed background files.
The exampel I used ot test this earlier used he background file
acis7sD2009-09-21bkgrndN0001.fits
where the N0001 in the file name is a version number. That older version uses the
 >i2 type and I had made marx to match that. In the current CALDB, there is a newer version of this file
acis7sD2009-09-21bkgrndN0002.fits
which is version N0002, which uses >i4 data datatype for PHA and PI columns.
My scripts did not catch that change, because I had the filename of the datafile hardcoded.
This commit switches marx2fits to use >i4 to match the event files (and background files)
that the current version of CIAO produces.