Add contributed user model
This model was developed by @andrea-belfiore for one specific application.
It's included in MARX not necessarily because we expect other users to need
exactly this model, but like the other examples, because it is a good case
to see how different, realistic types of user model can be implemented.

Thanks to @andrea-belfiore for sharing!