remove marx2img and marx2dpde
These tools are obsolete. The dpde format is not used any longer and the
marx2img tool should be replaces by marx2fits and "dmcopy" (or similar)
tools from CIAO.
By using CIAO tools directly, the header information,
datatypes etc. are the same as for observed data,
making it easier to combine both.

fix forgotten lines in makefile.in