Improve error message in marxasp
Previously, marxasp issues an error prompting the user to re-run the simulation
with DitherModel=INTERNAL, when run a simulation that used an ASPSOL input file.
 However, that's not what the typical user wants:
If the simulation was run with a real ASPSOL file, then that file should be used.
It's still true that marxasp should issue an error (since it won't generate anything),
but the message is more helpful now.

fixes #53