Remove IXO
Marx includes code to run simulations for IXO.
The reason for that is historic - it was used for perfomacne predictiosn in the ealry days of IXO.
However, it is clear that IXO will not be contoinued in the form simulated here and
new development of new missions will require nex code (or a difference codebase alltogether) anyway.
Thus, the chance of this code ever being used a gain is very low.
On the other hand, there is a (small, but not none) maintainance cost
to keep all the IXO mocduls for maing detectors etc. around.
Thus, this comit removes IXO code from the marx codebase.