Propagate changes to binary search
This is a follow-up on commit 10b0a24af2e26defce9b4e744f9f6ced59927b6b
That commit fixed a bug in JDMbinary_search_f. The same fix has to be applied in JDMbinary_search_d.
Also that fix changes the return values of the binary search routines to return the upper, not the lower index of a bin and so all parts of the marx code that use these routines need to be modified accordingly.