Remove EA_MIRROR
For a long time, the docs have said:
> This simple model is of historical interest only and is no longer accurate or supported. This option should not be selected.

This PR removes the EA mirror model.
It reduces maintenance (though small, it's not zero) and I find that any option that exists will be used by a user at some point and then they get inaccurate results. It's safer to remove an option that expressively should not be used any longer.

closes #75